Perform repeat logic in compile phase instead. (#10)
This allows us to provide support for recursive rendering of items and categories. In order to support this, we're making a few changes to the API:
- The app maker no longer provides their own ng-repeat directive. We do this internally now. We provide access to a "suggestion" on the scope of a suggestion item which gives them access to their data.
- When rendering categories, we're now requiring a specific data format. Any sub-children of a category must be in a `children` array, and the data passed to the source function itself must also be an array of objects.
- The suggestions should be an ordered list, so only Arrays are possible.

It'll now determine which item to highlight based on the suggestion item in the array as opposed to the HTML element in the Node list.

I've also stopped using lodash in the example to make @onlywei happy. Doing the categorizing loops by hand instead.