Fix keyboard highlight interaction by making a flattened UI array
Also add a new &-bound callback binding on ngc-omnibox called `isSelectable`. `isSelectable` has access to a `suggestion` value in its expression which is a single suggesiton object. If this expression returns true, then an item can be interacted with via keyboard and mouse. If it returns false, it cannot.

This can be used to disable selecting cateogry he aders, as well as just figuring out if individual items should be disabled.