Fix keyboard highlight interaction by making a flattened UI array (#12)
Also add a new &-bound callback binding on ngc-omnibox called `isSelectable`. `isSelectable` has access to a `suggestion` value in its expression which is a single suggesiton object. If this expression returns true, then an item can be interacted with via keyboard and mouse. If it returns false, it cannot.

This can be used to disable selecting category headers, as well as just figuring out if individual items should be disabled.

Also creates a function for un-highlighting. This might be useful in the future if we want to fire an event when this happens.