Remove the template from the Omnibox component (#18)
This allows the app maker to completely control the markup around the individual components so that they can better control the layout. Since we still want control over the input field, this requires adding a new directive for the input field that the app maker can use to decide where that element should be rendered in their markup.