Re-work how highlighting a choice works
In order to support additional methods other than TAB to focus on a choice, we have to simulate a choice having focus. Instead, the component itself has focus and listens for all keyboard events, then sets the state of the current choice that is highlighted. This allows us to use the arrow keys to focus on choices.

Highlighting a choice via the keyboard works in the following scenarios:
- The left arrow key (and backspace key) will select the last choice when the caret is at the beginning of the field.
- The right arrow key is will select the first choice when the caret is at the end of the field
- Tabbing always selects and blurs the field itself, and never the choices (unless a suggestion is highlighted, in which case it'll choose that suggestion and focus the field)