Fix digest cycle race conditions when updating suggestions (#33)
Depending on the app and how quickly your markup would get updated, sometimes when you re-opened the suggestions, you would see the old suggestions for a second until the markup was updated by Angular's digest cycle.

To combat this, we're now wrapping the contents of the suggestions directive in a wrapper so that we can use ng-if, which does a better job of syncing up the digest cycle. Since the user has control over the type of element and attributes that are applied to the suggestions directive, we also will try and honor their markup as much as possible, grabbing all the attributes from the now root suggestions element and applying them to the wrapper.