Clear the current suggestions while loading new ones (#53)
This fixes a confusing UI issue where if the suggestions load in slowly, you'll see the old suggestions for a while before the new ones pop in. We only clear out suggestions when hiding the panel. Otherwise if it's already open and you keep typing it blinks empty then shows new suggestions again, which is weird.