Call updateSuggestions w/o query on input change
As long as `canShowSuggestions()` is still cool with it.

This removes the assumption that the implementor doesn't want to do
something with the suggestions even when there's no query. They
might want to show something like instructions or helper UI, or even
some pre-generated default suggestions. This is now all possible.