Change the way that a CSS rule is added when component is focused
Change this logic so that instead of appending this rule to whatever the last stylesheet
exists in document.styleSheets, create a new stylesheet with this new rule in it.

This prevents a bug where if the last loaded stylesheet is from an external resource,
it would not allow access to the rules, and throw an error.