Add angular event propagation for onChosen handlers
This adds propagation for the angular events that trigger choose() calls
Unfortunately the $event name is taken by the custom event wrapper
already in use. It's attached to those events under the `sourceEvent`
name.

Once released this will allow `onChosen()` handlers to have access to
the event that triggered it's call